Move test fixtures as close as possible to their use
