+++ tests/conftest.py
@@ -252,6 +252,93 @@ def pytest_sessionfinish(
         )
 
 
+def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
+    """Return a parameterized entity."""
+    if "parameterized_entity" not in metafunc.fixturenames:
+        return
+
+    from copy import deepcopy
+    from pathlib import Path
+
+    import yaml
+
+    def get_version_name(uri: str) -> tuple[str, str]:
+        """Return the version and name part of a uri."""
+        import re
+
+        from dlite_entities_service.service.config import CONFIG
+
+        # namespace = "http://onto-ns.com/meta"
+        namespace = str(CONFIG.base_url).rstrip("/")
+
+        match = re.match(
+            rf"^{re.escape(namespace)}/(?P<version>[^/]+)/(?P<name>[^/]+)$", uri
+        )
+        assert match is not None, (
+            f"Could not retrieve version and name from {uri!r}. "
+            "URI must be of the form: "
+            f"{namespace}/{{version}}/{{name}}\n\n"
+            "Hint: Did you (inadvertently) set the base_url to something?"
+        )
+
+        return match.group("version") or "", match.group("name") or ""
+
+    def get_uri(entity: dict[str, Any]) -> str:
+        """Return the uri for an entity."""
+        namespace = entity.get("namespace")
+        version = entity.get("version")
+        name = entity.get("name")
+
+        assert not any(
+            _ is None for _ in (namespace, version, name)
+        ), "Could not retrieve namespace, version, and/or name from test entities."
+
+        return f"{namespace}/{version}/{name}"
+
+    results: list[ParameterizeGetEntities] = []
+
+    static_dir = (Path(__file__).parent / "static").resolve()
+
+    entities: list[dict[str, Any]] = yaml.safe_load(
+        (static_dir / "valid_entities.yaml").read_text()
+    )
+
+    for entity in entities:
+        uri = entity.get("uri") or get_uri(entity)
+
+        version, name = get_version_name(uri)
+
+        # Replace $ref with ref
+        backend_entity = deepcopy(entity)
+
+        # SOFT5
+        if isinstance(backend_entity["properties"], list):
+            backend_entity["properties"] = [
+                {key.replace("$ref", "ref"): value for key, value in property_.items()}
+                for property_ in backend_entity["properties"]
+            ]
+
+        # SOFT7
+        else:
+            for property_name, property_value in list(
+                backend_entity["properties"].items()
+            ):
+                backend_entity["properties"][property_name] = {
+                    key.replace("$ref", "ref"): value
+                    for key, value in property_value.items()
+                }
+
+        results.append(
+            ParameterizeGetEntities(entity, version, name, uri, backend_entity)
+        )
+
+    metafunc.parametrize(
+        "parameterized_entity",
+        results,
+        ids=[f"{_.version}/{_.name}" for _ in results],
+    )
+
+
 ## Pytest fixtures ##
 
 
@@ -476,184 +563,3 @@ def _mock_lifespan(live_backend: bool, monkeypatch: pytest.MonkeyPatch) -> None:
         monkeypatch.setattr(
             "dlite_entities_service.service.backend.clear_caches", lambda: None
         )
-
-
-@pytest.fixture()
-def _empty_backend_collection(
-    live_backend: bool, get_backend_user: GetBackendUserFixture
-) -> None:
-    """Empty the backend collection."""
-    from dlite_entities_service.service.backend import get_backend
-
-    backend_settings = {}
-    if live_backend:
-        backend_user = get_backend_user("readWrite")
-        backend_settings = {
-            "mongo_username": backend_user["username"],
-            "mongo_password": backend_user["password"],
-        }
-
-    backend: MongoDBBackend = get_backend(settings=backend_settings)
-    backend._collection.delete_many({})
-    assert backend._collection.count_documents({}) == 0
-
-
-@pytest.fixture()
-def client(
-    live_backend: bool, get_backend_user: GetBackendUserFixture
-) -> ClientFixture:
-    """Return the test client."""
-    import os
-
-    from fastapi.testclient import TestClient
-    from httpx import Client
-
-    from dlite_entities_service.main import APP
-    from dlite_entities_service.models.auth import Token
-    from dlite_entities_service.service.config import CONFIG
-
-    def _client(
-        auth_role: Literal["read", "readWrite"] | None = None
-    ) -> TestClient | Client:
-        """Return the test client with the given authentication role."""
-        if not live_backend:
-            return TestClient(
-                app=APP,
-                base_url=str(CONFIG.base_url),
-            )
-
-        if auth_role is None:
-            auth_role = "read"
-
-        assert auth_role in ("read", "readWrite"), (
-            f"Invalid authentication role {auth_role!r}. Must be either 'read' or "
-            "'readWrite'."
-        )
-
-        host, port = os.getenv("ENTITY_SERVICE_HOST", "localhost"), os.getenv(
-            "ENTITY_SERVICE_PORT", "8000"
-        )
-
-        base_url = f"http://{host}"
-
-        if port:
-            base_url += f":{port}"
-
-        backend_user = get_backend_user(auth_role)
-
-        with Client(base_url=base_url) as temp_client:
-            response = temp_client.post(
-                "/_auth/token",
-                data={
-                    "grant_type": "password",
-                    "username": backend_user["username"],
-                    "password": backend_user["password"],
-                },
-            )
-
-        assert response.is_success, response.text
-        try:
-            token = Token(**response.json())
-        except Exception as exc:
-            raise ValueError(
-                "Could not parse the response from the token endpoint. "
-                f"Response:\n{response.text}"
-            ) from exc
-
-        authentication_header = {
-            "Authorization": f"{token.token_type} {token.access_token}"
-        }
-
-        return Client(
-            base_url=f"http://{host}:{port}",
-            headers=authentication_header,
-        )
-
-    return _client
-
-
-def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
-    """Return a parameterized entity."""
-    if "parameterized_entity" not in metafunc.fixturenames:
-        return
-
-    from copy import deepcopy
-    from pathlib import Path
-
-    import yaml
-
-    def get_version_name(uri: str) -> tuple[str, str]:
-        """Return the version and name part of a uri."""
-        import re
-
-        from dlite_entities_service.service.config import CONFIG
-
-        # namespace = "http://onto-ns.com/meta"
-        namespace = str(CONFIG.base_url).rstrip("/")
-
-        match = re.match(
-            rf"^{re.escape(namespace)}/(?P<version>[^/]+)/(?P<name>[^/]+)$", uri
-        )
-        assert match is not None, (
-            f"Could not retrieve version and name from {uri!r}. "
-            "URI must be of the form: "
-            f"{namespace}/{{version}}/{{name}}\n\n"
-            "Hint: Did you (inadvertently) set the base_url to something?"
-        )
-
-        return match.group("version") or "", match.group("name") or ""
-
-    def get_uri(entity: dict[str, Any]) -> str:
-        """Return the uri for an entity."""
-        namespace = entity.get("namespace")
-        version = entity.get("version")
-        name = entity.get("name")
-
-        assert not any(
-            _ is None for _ in (namespace, version, name)
-        ), "Could not retrieve namespace, version, and/or name from test entities."
-
-        return f"{namespace}/{version}/{name}"
-
-    results: list[ParameterizeGetEntities] = []
-
-    static_dir = (Path(__file__).parent / "static").resolve()
-
-    entities: list[dict[str, Any]] = yaml.safe_load(
-        (static_dir / "valid_entities.yaml").read_text()
-    )
-
-    for entity in entities:
-        uri = entity.get("uri") or get_uri(entity)
-
-        version, name = get_version_name(uri)
-
-        # Replace $ref with ref
-        backend_entity = deepcopy(entity)
-
-        # SOFT5
-        if isinstance(backend_entity["properties"], list):
-            backend_entity["properties"] = [
-                {key.replace("$ref", "ref"): value for key, value in property_.items()}
-                for property_ in backend_entity["properties"]
-            ]
-
-        # SOFT7
-        else:
-            for property_name, property_value in list(
-                backend_entity["properties"].items()
-            ):
-                backend_entity["properties"][property_name] = {
-                    key.replace("$ref", "ref"): value
-                    for key, value in property_value.items()
-                }
-
-        results.append(
-            ParameterizeGetEntities(entity, version, name, uri, backend_entity)
-        )
-
-    metafunc.parametrize(
-        "parameterized_entity",
-        results,
-        ids=[f"{_.version}/{_.name}" for _ in results],
-    )
